+++ src/web_server/albumart.c
@@ -54,7 +54,7 @@ void webserver_send_albumart_redirect(struct mg_connection *nc, sds data) {
         sds redirect_uri = size == ALBUMART_THUMBNAIL
             ? sdscatfmt(sdsempty(),"/albumart-thumb?offset=0&uri=")
             : sdscatfmt(sdsempty(),"/albumart?offset=0&uri=");
-        redirect_uri = sds_urlencode(redirect_uri, uri, sdslen(uri));
+        redirect_uri = sds_urlencode(redirect_uri, uri, sdslen(uri), false);
         MYMPD_LOG_DEBUG(NULL, "Sending redirect to: %s", redirect_uri);
         webserver_send_header_found(nc, redirect_uri);
         FREE_SDS(redirect_uri);
@@ -162,7 +162,7 @@ bool request_handler_albumart_by_uri(struct mg_connection *nc, struct mg_http_me
                     //full uri, send redirect to covercache proxy
                     //use relative path to support hosting myMPD behind a reverse proxy in a subdir
                     sds redirect_uri = sdsnew("proxy-covercache?uri=");
-                    redirect_uri = sds_urlencode(redirect_uri, extimg, sdslen(extimg));
+                    redirect_uri = sds_urlencode(redirect_uri, extimg, sdslen(extimg), false);
                     webserver_send_header_found(nc, redirect_uri);
                     FREE_SDS(redirect_uri);
                     FREE_SDS(uri);